Delete Route
fixed deleting a comment
