+++ Picturest/src/main/java/com/codingdojo/picturest/controllers/PhotosController.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 
 import com.codingdojo.picturest.models.Photo;
+import com.codingdojo.picturest.services.CommentService;
 import com.codingdojo.picturest.services.PhotoService;
 import com.codingdojo.picturest.services.UserService;
 
@@ -27,6 +28,9 @@ public class PhotosController {
 	//bringing in user service for liking and commenting on photo
 	@Autowired
 	private UserService userService;
+	
+	@Autowired
+	private CommentService commentService;
 	
 	
 // ------------------------- SAVE PHOTO TO DB ------------------------- //
@@ -48,12 +52,12 @@ public class PhotosController {
 		model.addAttribute("allLikes", photoService.showOnePhoto(photoId).getUsersWhoLikePhoto());
 		
 		//add all of the comments on the photo - get all comments by photoId?
-		model.addAttribute("allCommentsByPhotoId", userService.getCommentsByPhotoId(photoId));
+		model.addAttribute("allCommentsByPhotoId", commentService.getCommentsByPhotoId(photoId));
 		return "showOnePicture.jsp";
 		
 	}
 	
-	
+
 // ------------------------- EDIT ONE PHOTO ------------------------- //
 	// TAKE USER TO EDIT PHOTO PAGE. NEED MODEL ATTRIBUTE FOR THE FORM:FORM BIND INFORMATION
 	@GetMapping("/edit/photo/{id}")
